move filter logic to be used with batch analysis
