+++ src/app/motor-inventory/batch-analysis/batch-analysis-table/batch-analysis-table.component.css
@@ -1,0 +1,36 @@
+.sticky-left{
+    position: -webkit-sticky;
+    position: sticky;
+    left: 0;
+}
+
+.table{
+    font-size: small;
+    margin-bottom: 0px;
+}
+
+
+.name-cell, .header-cell{
+    background-color: white;
+}
+
+.name-cell{
+    font-weight: bold;
+}
+
+th, td{
+    min-width: 75px;
+}
+
+.max-height{
+    max-height: 100%;
+}
+
+.sort-by-active{
+    text-decoration: underline;
+}
+
+.header-cell:hover{
+    text-decoration: underline;
+    cursor: pointer;
+}
