+++ src/app/motor-inventory/batch-analysis/batch-analysis-table/batch-analysis-table.component.html
@@ -1,0 +1,164 @@
+<table class="table table-sm table-bordered table-hover">
+    <thead>
+        <tr>
+            <th class="header-cell">
+                Motor Name
+            </th>
+            <th class="header-cell">
+                Department
+            </th>
+            <th class="header-cell">
+                Operating Hours
+            </th>
+            <th class="header-cell">
+                Motor Size
+            </th>
+            <th class="header-cell">
+                Average Load Factor
+            </th>
+            <th class="header-cell">
+                Current Efficiency
+            </th>
+            <!-- <th class="header-cell">
+                Rewind Efficiency Loss
+            </th> -->
+            <th class="header-cell">
+                Rewind Cost
+            </th>
+            <th class="header-cell">
+                Replacement Efficiency
+            </th>
+            <th class="header-cell">
+                Replacement Purchase Cost
+            </th>
+            <th class="header-cell">
+                Current Energy Use
+            </th>
+            <th class="header-cell">
+                Current Energy Cost
+            </th>
+            <th class="header-cell">
+                Rewind Energy Use
+            </th>
+            <th class="header-cell">
+                Rewind Energy Cost
+            </th>
+            <th class="header-cell">
+                Replacement Energy Use
+            </th>
+            <th class="header-cell">
+                Replacement Energy Cost
+            </th>
+            <th class="header-cell">
+                Replacement Now Payback
+            </th>
+            <th class="header-cell">
+                Replacement Fail Payback
+            </th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr *ngFor="let batchAnalysisDataItem of batchAnalysisDataItems">
+            <td class="sticky-left name-cell">
+                {{batchAnalysisDataItem.motorName}}
+            </td>
+            <td>
+                {{batchAnalysisDataItem.departmentName}}
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.operatingHours">
+                    {{batchAnalysisDataItem.operatingHours}} hrs/yr
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.operatingHours">
+                    &mdash;
+                </span>
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.motorSize">
+                    {{batchAnalysisDataItem.motorSize}} hp
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.motorSize">
+                    &mdash;
+                </span>
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.averageLoadFactor">
+                    {{batchAnalysisDataItem.averageLoadFactor | number:'1.0-1'}} %
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.averageLoadFactor">
+                    &mdash;
+                </span>
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.currentEfficiency">
+                    {{batchAnalysisDataItem.currentEfficiency | number:'1.0-1'}} %
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.currentEfficiency">
+                    &mdash;
+                </span>
+            </td>
+            <!-- <td>
+                <span *ngIf="batchAnalysisDataItem.rewindEfficiencyLoss">
+                    {{batchAnalysisDataItem.rewindEfficiencyLoss | number:'1.0-1'}} %
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.rewindEfficiencyLoss">
+                    &mdash;
+                </span>
+            </td> -->
+            <td>
+                <span *ngIf="batchAnalysisDataItem.rewindCost">
+                    {{batchAnalysisDataItem.rewindCost | currency}}
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.rewindCost">
+                    &mdash;
+                </span>
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.replacementEfficiency">
+                    {{batchAnalysisDataItem.replacementEfficiency | number:'1.0-1'}} %
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.replacementEfficiency">
+                    &mdash;
+                </span>
+            </td>
+            <td>
+                <span *ngIf="batchAnalysisDataItem.replacementPurchaseCost">
+                    {{batchAnalysisDataItem.replacementPurchaseCost | currency }}
+                </span>
+                <span *ngIf="!batchAnalysisDataItem.replacementPurchaseCost">
+                    &mdash;
+                </span>
+            </td>
+            <td colspan="8" *ngIf="!batchAnalysisDataItem.isBatchAnalysisValid">
+                Unable to run batch analysis on motor, data missing:
+                <span class="bold" *ngFor="let missing of batchAnalysisDataItem.missingData; let last=last"> {{missing}}<span *ngIf="!last">,</span><span *ngIf="last">.</span>
+                </span>
+                Update the motor item with the proper data to run batch analysis.
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.currentEnergyUse | number:'1.0-1'}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.currentEnergyCost | currency}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.rewindEnergyUse | number:'1.0-1'}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.rewindEnergyCost | currency}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.replacementEnergyUse | number:'1.0-1'}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.replacementEnergyCost | currency}}
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.replacementNowPayback | number:'1.0-1'}} yrs
+            </td>
+            <td *ngIf="batchAnalysisDataItem.isBatchAnalysisValid">
+                {{batchAnalysisDataItem.replacementFailPayback | number:'1.0-1'}} yrs
+            </td>
+        </tr>
+    </tbody>
+</table>
