+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-graphs/inventory-summary-graphs-menu/inventory-summary-graphs-menu.component.ts
@@ -25,6 +25,7 @@ export class InventorySummaryGraphsMenuComponent implements OnInit {
   constructor(private motorInventoryService: MotorInventoryService, private inventorySummaryGraphService: InventorySummaryGraphsService, private motorInventorySummaryService: MotorInventorySummaryService) { }
 
   ngOnInit(): void {
+    this.setOptions();
     this.selectedFieldSub = this.inventorySummaryGraphService.selectedField.subscribe(val => {
       if (val) {
         this.selectedField = val;
