+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-table/inventory-summary-table.component.html
@@ -2,7 +2,7 @@
     <table class="table table-sm table-bordered table-hover">
         <thead>
             <tr>
-                <th class="header-cell sticky-top" *ngFor="let field of inventorySummaryData.fields; let index = index;"
+                <th class="header-cell" *ngFor="let field of inventorySummaryData.fields; let index = index;"
                     [ngClass]="{'sticky-left': index == 0}"><a
                         [ngClass]="{'sort-by-active': field.value == sortByField}"
                         (click)="setSortByField(field.value)">
